Extract workspace gallery actions hook
