+++ tests/workspace-gallery-rail-contract.test.ts
@@ -47,10 +47,6 @@ test('composite preview modal button opens direct preview groups', () => {
 });
 
 test('composite preview preserves preview urls but plays canonical video urls', () => {
-  const appSource = fs.readFileSync(
-    path.join(process.cwd(), 'frontend/app/(core)/(workspace)/app/AppClient.tsx'),
-    'utf8'
-  );
   const renderGroupSource = fs.readFileSync(
     path.join(process.cwd(), 'frontend/app/(core)/(workspace)/app/_lib/workspace-render-groups.ts'),
     'utf8'
@@ -59,12 +55,16 @@ test('composite preview preserves preview urls but plays canonical video urls', 
     path.join(process.cwd(), 'frontend/app/(core)/(workspace)/app/_lib/workspace-video-settings.ts'),
     'utf8'
   );
+  const galleryActionsHookSource = fs.readFileSync(
+    path.join(process.cwd(), 'frontend/app/(core)/(workspace)/app/_hooks/useWorkspaceGalleryActions.ts'),
+    'utf8'
+  );
   const dockSource = fs.readFileSync(
     path.join(process.cwd(), 'frontend/components/groups/CompositePreviewDock.tsx'),
     'utf8'
   );
 
-  assert.match(appSource, /previewVideoUrl:\s*tile\.previewVideoUrl/);
+  assert.match(galleryActionsHookSource, /previewVideoUrl:\s*tile\.previewVideoUrl/);
   assert.match(videoSettingsSource, /previewVideoUrl:\s*patch\.previewVideoUrl\s*\?\?\s*current\.previewVideoUrl/);
   assert.match(videoSettingsSource, /previewUrl:\s*patch\.previewVideoUrl\s*\?\?\s*item\.previewUrl/);
   assert.match(renderGroupSource, /previewVideoUrl:\s*gatingActive\s*\?\s*null\s*:\s*item\.previewVideoUrl\s*\?\?\s*null/);
@@ -77,15 +77,19 @@ test('rail history previews do not mutate the active render group', () => {
     path.join(process.cwd(), 'frontend/app/(core)/(workspace)/app/AppClient.tsx'),
     'utf8'
   );
+  const galleryActionsHookSource = fs.readFileSync(
+    path.join(process.cwd(), 'frontend/app/(core)/(workspace)/app/_hooks/useWorkspaceGalleryActions.ts'),
+    'utf8'
+  );
 
-  assert.match(appSource, /if\s*\(group\.source === 'active'\)\s*\{\s*setActiveGroupId\(group\.id\)/);
-  assert.match(appSource, /const openGroupViaGallery = useCallback\(\s*\(\s*group: GroupSummary\s*\)\s*=>\s*\{\s*handleGalleryGroupAction\(group,\s*'open',\s*\{\s*autoPlayPreview:\s*true\s*\}\);/);
-  assert.doesNotMatch(appSource, /const openGroupViaGallery[\s\S]*?setActiveGroupId\(group\.id\)[\s\S]*?\[handleGalleryGroupAction\]/);
-  const activeOpenStart = appSource.indexOf("if (group.source === 'active')");
-  const activeOpenEnd = appSource.indexOf("if (action === 'continue')", activeOpenStart);
+  assert.match(galleryActionsHookSource, /if\s*\(group\.source === 'active'\)\s*\{\s*setActiveGroupId\(group\.id\)/);
+  assert.match(galleryActionsHookSource, /const openGroupViaGallery = useCallback\(\s*\(\s*group: GroupSummary\s*\)\s*=>\s*\{\s*handleGalleryGroupAction\(group,\s*'open',\s*\{\s*autoPlayPreview:\s*true\s*\}\);/);
+  assert.doesNotMatch(galleryActionsHookSource, /const openGroupViaGallery[\s\S]*?setActiveGroupId\(group\.id\)[\s\S]*?\[handleGalleryGroupAction\]/);
+  const activeOpenStart = galleryActionsHookSource.indexOf("if (group.source === 'active')");
+  const activeOpenEnd = galleryActionsHookSource.indexOf("if (action === 'continue')", activeOpenStart);
   assert.notEqual(activeOpenStart, -1);
   assert.notEqual(activeOpenEnd, -1);
-  const activeOpenBlock = appSource.slice(activeOpenStart, activeOpenEnd);
+  const activeOpenBlock = galleryActionsHookSource.slice(activeOpenStart, activeOpenEnd);
   assert.doesNotMatch(activeOpenBlock, /applyVideoSettingsFromTile\(tile\);/);
   assert.doesNotMatch(activeOpenBlock, /hydrateVideoSettingsFromJob\(heroJobId\);/);
   assert.doesNotMatch(
